Split CreditCode/Reward repos out, added commander
Moved CreditCode/Reward into payments
Added a CreditRewardCommander
