+++ kifi-backend/modules/shoebox/app/com/keepit/payments/AccountEvent.scala
@@ -7,7 +7,7 @@ import com.keepit.common.crypto.{ ModelWithPublicId, ModelWithPublicIdCompanion 
 import com.keepit.common.db.{ Id, ModelWithState, State, States }
 import com.keepit.common.mail.EmailAddress
 import com.keepit.common.time._
-import com.keepit.model.{ CreditReward, User }
+import com.keepit.model.User
 import com.keepit.common.mail.EmailAddress
 import com.kifi.macros.json
 import org.joda.time.DateTime
